detox-collective/backend/requirements.txt
Flask==2.3.4
firebase-admin==6.1.2
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.4.0
requests==2.32.0
